Revert "[Tracer] Rejit refactor to support multiple products in the same method (#5533)"
This reverts commit 61603e1639aace563c3b1f950c1c2f083b2d03c0.
