+++ tracer/src/Datadog.Tracer.Native/cor_profiler.cpp
@@ -358,7 +358,7 @@ HRESULT STDMETHODCALLTYPE CorProfiler::Initialize(IUnknown* cor_profiler_info_un
 
     if (isIastEnabled || isRaspEnabled)
     {
-        _dataflow = new iast::Dataflow(info_, rejit_handler);
+        _dataflow = new iast::Dataflow(info_);
         if (FAILED(_dataflow->Init()))
         {
             Logger::Error("Callsite Dataflow failed to initialize");
@@ -1297,13 +1297,13 @@ HRESULT STDMETHODCALLTYPE CorProfiler::Shutdown()
     // to prevent it from unloading while in use
     auto modules = module_ids.Get();
 
+    DEL(_dataflow)
+
     if (rejit_handler != nullptr)
     {
         rejit_handler->Shutdown();
         rejit_handler = nullptr;
     }
-
-    DEL(_dataflow);
 
     auto definitions = definitions_ids.Get();
 
@@ -1821,6 +1821,19 @@ void CorProfiler::InternalAddInstrumentation(WCHAR* id, CallTargetDefinition* it
             {
                 integration_definitions_.push_back(integration);
             }
+
+            Logger::Info("Total number of modules to analyze: ", modules->size());
+            if (rejit_handler != nullptr)
+            {
+                auto promise = std::make_shared<std::promise<ULONG>>();
+                std::future<ULONG> future = promise->get_future();
+                tracer_integration_preprocessor->EnqueueRequestRejitForLoadedModules(modules.Ref(),
+                    integrationDefinitions, promise);
+
+                // wait and get the value from the future<int>
+                const auto& numReJITs = future.get();
+                Logger::Debug("Total number of ReJIT Requested: ", numReJITs);
+            }
         }
         else
         {
@@ -1835,19 +1848,19 @@ void CorProfiler::InternalAddInstrumentation(WCHAR* id, CallTargetDefinition* it
                     integration_definitions_.push_back(integration);
                 }
             }
-        }
 
-        Logger::Info("Total number of modules to analyze: ", modules->size());
-        if (rejit_handler != nullptr)
-        {
-            auto promise = std::make_shared<std::promise<ULONG>>();
-            std::future<ULONG> future = promise->get_future();
-            tracer_integration_preprocessor->EnqueueRequestRejitForLoadedModules(modules.Ref(), integrationDefinitions,
-                                                                                 promise);
+            Logger::Info("Total number of modules to analyze: ", modules->size());
+            if (rejit_handler != nullptr)
+            {
+                auto promise = std::make_shared<std::promise<ULONG>>();
+                std::future<ULONG> future = promise->get_future();
+                tracer_integration_preprocessor->EnqueueRequestRevertForLoadedModules(modules.Ref(),
+                    integrationDefinitions, promise);
 
-            // wait and get the value from the future<int>
-            const auto& numReJITs = future.get();
-            Logger::Debug("Total number of ReJIT Requested: ", numReJITs);
+                // wait and get the value from the future<int>
+                const auto& numReJITs = future.get();
+                Logger::Debug("Total number of Revert Requested: ", numReJITs);
+            }
         }
 
         Logger::Info("InitializeProfiler: Total integrations in profiler: ", integration_definitions_.size());
@@ -2003,7 +2016,8 @@ long CorProfiler::DisableCallTargetDefinitions(UINT32 disabledCategories)
             auto modules = module_ids.Get();
             auto promise = std::make_shared<std::promise<ULONG>>();
             std::future<ULONG> future = promise->get_future();
-            tracer_integration_preprocessor->EnqueueRequestRejitForLoadedModules(modules.Ref(), affectedDefinitions, promise);
+            tracer_integration_preprocessor->EnqueueRequestRevertForLoadedModules(modules.Ref(), affectedDefinitions,
+                                                                                  promise);
 
             // wait and get the value from the future<int>
             numReverts = future.get();
