+++ tracer/src/Datadog.Tracer.Native/debugger_method_rewriter.cpp
@@ -261,9 +261,7 @@ HRESULT DebuggerMethodRewriter::LoadInstanceIntoStack(FunctionInfo* caller, bool
     return S_OK;
 }
 
-HRESULT DebuggerMethodRewriter::Rewrite(RejitHandlerModule* moduleHandler, RejitHandlerModuleMethod* methodHandler,
-                                        ICorProfilerFunctionControl* pFunctionControl,
-                                        ICorProfilerInfo* pCorProfilerInfo)
+HRESULT DebuggerMethodRewriter::Rewrite(RejitHandlerModule* moduleHandler, RejitHandlerModuleMethod* methodHandler, ICorProfilerFunctionControl* pFunctionControl)
 {
     const auto moduleId = moduleHandler->GetModuleId();
     const auto methodId = methodHandler->GetMethodDef();
@@ -333,8 +331,7 @@ HRESULT DebuggerMethodRewriter::Rewrite(RejitHandlerModule* moduleHandler, Rejit
         Logger::Info("Applying ", methodProbes.size(), " method probes, ", lineProbes.size(), " line probes and ",
                      spanOnMethodProbes.size(), " span probes on methodDef: ", methodHandler->GetMethodDef());
 
-        auto hr = Rewrite(moduleHandler, methodHandler, pFunctionControl, pCorProfilerInfo, methodProbes, lineProbes,
-                          spanOnMethodProbes);
+        auto hr = Rewrite(moduleHandler, methodHandler, pFunctionControl, methodProbes, lineProbes, spanOnMethodProbes);
 
         if (hr == S_OK)
         {
@@ -2111,7 +2108,6 @@ void DebuggerMethodRewriter::MarkAllSpanOnMethodProbesAsError(SpanProbeOnMethodD
 HRESULT DebuggerMethodRewriter::Rewrite(RejitHandlerModule* moduleHandler,
                                         RejitHandlerModuleMethod* methodHandler,
                                         ICorProfilerFunctionControl* pFunctionControl,
-                                        ICorProfilerInfo* pCorProfilerInfo,
                                         MethodProbeDefinitions& methodProbes,
                                         LineProbeDefinitions& lineProbes,
                                         SpanProbeOnMethodDefinitions& spanOnMethodProbes) const
@@ -2161,7 +2157,7 @@ HRESULT DebuggerMethodRewriter::Rewrite(RejitHandlerModule* moduleHandler,
     }
 
     // *** Create rewriter
-    ILRewriter rewriter(pCorProfilerInfo, pFunctionControl, module_id, function_token);
+    ILRewriter rewriter(m_corProfiler->info_, pFunctionControl, module_id, function_token);
     auto hr = rewriter.Import();
     if (FAILED(hr))
     {
