+++ tracer/src/Datadog.Tracer.Native/debugger_rejit_preprocessor.cpp
@@ -11,13 +11,6 @@ namespace debugger
 
 // DebuggerRejitPreprocessor
 
-DebuggerRejitPreprocessor::DebuggerRejitPreprocessor(CorProfiler* corProfiler,
-                                                     std::shared_ptr<RejitHandler> rejit_handler,
-                                                     std::shared_ptr<RejitWorkOffloader> work_offloader) :
-    RejitPreprocessor(corProfiler, rejit_handler, work_offloader, RejitterPriority::Critical)
-{
-}
-
 ULONG DebuggerRejitPreprocessor::PreprocessLineProbes(
     const std::vector<ModuleID>& modules, const std::vector<std::shared_ptr<LineProbeDefinition>>& lineProbes,
     std::vector<MethodIdentifier>& rejitRequests)
@@ -99,7 +92,7 @@ ULONG DebuggerRejitPreprocessor::PreprocessLineProbes(
                 continue;
             }
 
-            auto moduleHandler = GetOrAddModule(moduleInfo.id);
+            auto moduleHandler = m_rejit_handler->GetOrAddModule(moduleInfo.id);
             if (moduleHandler == nullptr)
             {
                 Logger::Warn("Module handler is null, this only happens if the RejitHandler has been shutdown.");
