+++ tracer/src/Datadog.Tracer.Native/fault_tolerant_cor_profiler_function_control.cpp
@@ -7,8 +7,7 @@
 using namespace fault_tolerant;
 
 FaultTolerantCorProfilerFunctionControl::FaultTolerantCorProfilerFunctionControl(
-    ICorProfilerFunctionControl* corProfilerFunctionControl, ICorProfilerInfo* corProfilerInfo, ModuleID moduleId,
-    mdMethodDef methodId,
+    ICorProfilerFunctionControl* corProfilerFunctionControl, ModuleID moduleId, mdMethodDef methodId,
     RejitHandlerModule* moduleHandler, RejitHandlerModuleMethod* methodHandler,
     InjectSuccessfulInstrumentationLambda injectSuccessfulInstrumentation) :
     m_moduleId(moduleId),
@@ -16,8 +15,7 @@ FaultTolerantCorProfilerFunctionControl::FaultTolerantCorProfilerFunctionControl
     moduleHandler(moduleHandler),
     methodHandler(methodHandler),
     injectSuccessfulInstrumentation(std::move(injectSuccessfulInstrumentation)),
-    m_pICorProfilerFunctionControl(corProfilerFunctionControl),
-    m_pCorProfilerInfo(corProfilerInfo)
+    m_pICorProfilerFunctionControl(corProfilerFunctionControl)
 {
 }
 
@@ -63,8 +61,7 @@ ULONG STDMETHODCALLTYPE FaultTolerantCorProfilerFunctionControl::Release()
 HRESULT FaultTolerantCorProfilerFunctionControl::SetILFunctionBody(ULONG cbNewILMethodHeader,
                                                                    LPCBYTE pbNewILMethodHeader)
 {
-    return injectSuccessfulInstrumentation(moduleHandler, methodHandler, m_pICorProfilerFunctionControl,
-                                           m_pCorProfilerInfo, pbNewILMethodHeader);
+    return injectSuccessfulInstrumentation(moduleHandler, methodHandler, m_pICorProfilerFunctionControl, pbNewILMethodHeader);
 }
 
 HRESULT FaultTolerantCorProfilerFunctionControl::SetCodegenFlags(DWORD flags)
