+++ tracer/src/Datadog.Tracer.Native/fault_tolerant_tracker.cpp
@@ -1,11 +1,20 @@
 #include "fault_tolerant_tracker.h"
 
+void fault_tolerant::FaultTolerantTracker::RequestRevert(ModuleID moduleId, mdMethodDef methodId, std::shared_ptr<RejitHandler> rejit_handler)
+{
+    std::vector<MethodIdentifier> requests = {{moduleId, methodId}};
+    auto promise = std::make_shared<std::promise<void>>();
+    auto future = promise->get_future();
+    rejit_handler->EnqueueRequestRevert(requests, promise);
+    future.get();
+}
+
 void fault_tolerant::FaultTolerantTracker::RequestRejit(ModuleID moduleId, mdMethodDef methodId, std::shared_ptr<RejitHandler> rejit_handler)
 {
     std::vector<MethodIdentifier> requests = {{moduleId, methodId}};
     auto promise = std::make_shared<std::promise<void>>();
     auto future = promise->get_future();
-    rejit_handler->EnqueueRequestRejit(requests, promise, true);
+    rejit_handler->EnqueueRequestRejit(requests, promise);
     future.get();
 }
 
@@ -120,6 +129,7 @@ void fault_tolerant::FaultTolerantTracker::AddSuccessfulInstrumentationId(
 {
     std::lock_guard lock(_successfulInstrumentationIdsMutex);
 
+    RequestRevert(moduleId, methodId, rejit_handler);
     RequestRejit(moduleId, methodId, rejit_handler);
 
     const auto methodIdentifier = trace::MethodIdentifier(moduleId, methodId);
@@ -155,7 +165,7 @@ bool fault_tolerant::FaultTolerantTracker::ShouldHeal(ModuleID moduleId, mdMetho
 
     if (shouldHeal)
     {
-        RequestRejit(moduleId, methodId, rejit_handler);
+        RequestRevert(moduleId, methodId, rejit_handler);
     }
 
     return shouldHeal;
