+++ tracer/src/Datadog.Tracer.Native/iast/dataflow.h
@@ -2,13 +2,6 @@
 #include "../../../../shared/src/native-src/pal.h"
 #include "iast_util.h"
 #include "aspect.h"
-#include "../rejit_handler.h"
-#include "../rejit_preprocessor.h"
-
-namespace trace
-{
-    class FunctionControlWrapper;
-}
 
 namespace iast
 {
@@ -23,7 +16,7 @@ namespace iast
     struct ILInstr;
     enum class DataflowAspectFilterValue;
     class AspectFilter;
-    struct DataflowContext;
+    struct InstrumentResult;
 
     struct RewriteMethodResponse
     {
@@ -47,12 +40,12 @@ namespace iast
         AspectFilter* GetFilter(DataflowAspectFilterValue filterValue);
     };
 
-    class Dataflow : public trace::Rejitter
+    class Dataflow
     {
         friend class ModuleInfo;
         friend class ModuleAspects;
     public:
-        Dataflow(ICorProfilerInfo* profiler, std::shared_ptr<RejitHandler> rejitHandler);
+        Dataflow(ICorProfilerInfo* profiler);
         virtual ~Dataflow();
     private:
         CS _cs;
@@ -72,6 +65,8 @@ namespace iast
         std::vector<WSTRING> _methodAttributeIncludeFilters;
         std::vector<WSTRING> _methodAttributeExcludeFilters;
 
+        bool _traceJitMethods = false;
+
     protected:
         bool _initialized = false;
         bool _loaded = false;
@@ -80,11 +75,11 @@ namespace iast
         std::vector<DataflowAspect*> _aspects;
         std::map<ModuleID, ModuleAspects*> _moduleAspects;
 
-        HRESULT RewriteMethod(MethodInfo* method, trace::FunctionControlWrapper* pFunctionControl = nullptr);
-        MethodInfo* JITProcessMethod(ModuleID moduleId, mdToken methodId, trace::FunctionControlWrapper* pFunctionControl = nullptr);
+        HRESULT RewriteMethod(MethodInfo* method, ICorProfilerFunctionControl* pFunctionControl);
+        MethodInfo* JITProcessMethod(ModuleID moduleId, mdToken methodId, bool isRejit = false);
 
         std::vector<DataflowAspectReference*> GetAspects(ModuleInfo* module);
-        static bool InstrumentInstruction(DataflowContext& context, std::vector<DataflowAspectReference*>& aspects);
+        static InstrumentResult InstrumentInstruction(ILRewriter* rewriter, ILInstr* instruction, std::vector<DataflowAspectReference*>& aspects);
 
     public:
         bool IsInitialized();
@@ -124,14 +119,5 @@ namespace iast
 
         bool IsInlineEnabled(ModuleID calleeModuleId, mdToken calleeMethodId);
         bool JITCompilationStarted(ModuleID moduleId, mdToken methodId);
-
-    public:
-        void Shutdown() override;
-        RejitHandlerModule* GetOrAddModule(ModuleID moduleId) override;
-        bool HasModuleAndMethod(ModuleID moduleId, mdMethodDef methodDef) override;
-        void RemoveModule(ModuleID moduleId) override;
-        void AddNGenInlinerModule(ModuleID moduleId) override;
-
-        HRESULT RejitMethod(trace::FunctionControlWrapper& functionControl) override;
     };
-} // namespace iast
+}
