+++ tracer/src/Datadog.Tracer.Native/iast/dataflow_il_rewriter.h
@@ -107,7 +107,7 @@ namespace iast
         ILInstr* NewILInstr(OPCODE opcode = CEE_COUNT, ULONG32 arg = 0);
         ILInstr* InsertBefore(ILInstr* pWhere, ILInstr* pWhat, bool updateReferences = true);
         ILInstr* InsertAfter(ILInstr* pWhere, ILInstr* pWhat, bool updateReferences = true);
-        HRESULT Export();
+        HRESULT Export(const std::string& applyMessage = "");
         ILInstr* GetILList();
         ILInstr* GetInstrFromOffset(unsigned offset);
         std::vector<EHClause*> GetExceptionHandlers();
