+++ tracer/src/Datadog.Tracer.Native/iast/hardcoded_secrets_method_analyzer.cpp
@@ -46,7 +46,7 @@ bool HardcodedSecretsMethodAnalyzer::ProcessMethod(MethodInfo* method)
                 userStrings.push_back(str);
             }
         }
-        hr = method->CommitILRewriter(true);
+        hr = method->CommitILRewriter();
     }
 
     if (userStrings.size() > 0)
