+++ tracer/src/Datadog.Tracer.Native/iast/method_info.cpp
@@ -266,7 +266,7 @@ namespace iast
         {
             methodKeyBuilder << " Processed ";
         }
-        if (HasChanges())
+        if (HasChanged())
         {
             methodKeyBuilder << " Changed ";
         }
@@ -289,15 +289,7 @@ namespace iast
     {
         _isProcessed = true;
     }
-    void MethodInfo::SetInstrumented(bool instrumented)
-    {
-        _isInstrumented = instrumented;
-    }
-    bool MethodInfo::IsInstrumented()
-    {
-        return _isInstrumented;
-    }
-    bool MethodInfo::HasChanges()
+    bool MethodInfo::HasChanged()
     {
         return _pMethodIL;
     }
@@ -311,33 +303,21 @@ namespace iast
     }
     bool MethodInfo::IsInlineEnabled()
     {
-        return _module->IsInlineEnabled() && !_disableInlining && !IsInstrumented();
+        return _module->IsInlineEnabled() && !_disableInlining && !HasChanged();
     }
     void MethodInfo::DisableInlining()
     {
         _disableInlining = true;
     }
 
-    HRESULT MethodInfo::GetILRewriter(ILRewriter** pRewriter, ICorProfilerInfo* pCorProfilerInfo)
+    HRESULT MethodInfo::GetILRewriter(ILRewriter** pRewriter)
     {
         HRESULT hr = S_FALSE;
         if (_rewriter == nullptr)
         {
             hr = S_OK;
             _rewriter = new ILRewriter(this);
-            if (!pCorProfilerInfo)
-            {
-                hr = _rewriter->Import();
-            }
-            else
-            {
-                LPCBYTE pMethodIL = nullptr;
-                hr = pCorProfilerInfo->GetILFunctionBody(_module->_id, _id, &pMethodIL, nullptr);
-                if (SUCCEEDED(hr))
-                {
-                    hr = _rewriter->Import(pMethodIL);
-                }
-            }
+            hr = _rewriter->Import();
             if (FAILED(hr))
             {
                 DEL(_rewriter);
@@ -346,12 +326,12 @@ namespace iast
         *pRewriter = _rewriter;
         return hr;
     }
-    HRESULT MethodInfo::CommitILRewriter(bool abort)
+    HRESULT MethodInfo::CommitILRewriter(const std::string& applyMessage)
     {
         HRESULT hr = S_FALSE;
-        if (_rewriter != nullptr && !abort)
+        if (_rewriter != nullptr)
         {
-            hr = _rewriter->Export();
+            hr = _rewriter->Export(applyMessage);
         }
         DEL(_rewriter);
         return hr;
@@ -378,6 +358,29 @@ namespace iast
         return hr;
     }
 
+    void MethodInfo::DumpIL(const std::string message, ULONG pnMethodIL, LPCBYTE pMethodIL)
+    {
+        if (!pMethodIL)
+        {
+            GetMethodIL(&pMethodIL, &pnMethodIL);
+        }
+        ILRewriter ilRewriter(this);
+        if (FAILED(ilRewriter.Import(pMethodIL)))
+        {
+            trace::Logger::Info("Dumping IL ", message, " : ", GetFullName(), " IL Verification FAILED ( Error on ILImport ) !!!");
+            return;
+        }
+
+        ILAnalysis analysis(&ilRewriter);
+        auto correct = analysis.IsStackValid();
+        auto verificationFail = analysis.GetError();
+        analysis.Dump(message);
+        if (!correct)
+        {
+            trace::Logger::Info("Dumping IL ", message, " : ", GetFullName(), " IL Verification FAILED ( ", verificationFail, " ) !!!");
+        }
+    }
+
     HRESULT MethodInfo::SetMethodIL(ULONG nSize, LPCBYTE pMethodIL, ICorProfilerFunctionControl* pFunctionControl)
     {
         bool isRejit = pFunctionControl != nullptr;
@@ -389,6 +392,12 @@ namespace iast
         {
             trace::Logger::Debug("Same function body detected. Skipping SetMethodIL ", GetKey());
             return S_FALSE;
+        }
+
+        if (!isRejit && _module->ExcludeInChaining())
+        {
+            DEL_ARR(pMethodIL);
+            return S_OK;
         }
 
         bool correct = true;
@@ -462,16 +471,8 @@ namespace iast
         HRESULT hr = S_OK;
         if (pFunctionControl)
         {
-            if (HasChanges())
-            {
-                hr = pFunctionControl->SetILFunctionBody(_nMethodIL, _pMethodIL);
-                trace::Logger::Debug("MethodInfo::ApplyFinalInstrumentation ReJIT from ", _nOriginalMehodIL, " to ",
-                                     _nMethodIL, " on ", GetKey(), " hr=", Hex(hr));
-            }
-            else
-            {
-                trace::Logger::Debug("MethodInfo::ApplyFinalInstrumentation ReJIT SKIPPED (method did not change)");
-            }
+            hr = pFunctionControl->SetILFunctionBody(_nMethodIL, _pMethodIL);
+            trace::Logger::Debug("MethodInfo::ApplyFinalInstrumentation ReJIT from ", _nOriginalMehodIL, " to ", _nMethodIL, " on ", GetKey(), " hr=", Hex(hr));
         }
         else
         {
@@ -483,7 +484,11 @@ namespace iast
                 }
                 return hr;
             }
-            if (!HasChanges())
+            if (_module->ExcludeInChaining())
+            {
+                return S_FALSE;
+            }
+            if (!HasChanged())
             {
                 trace::Logger::Error("ERROR: MethodInfo::ApplyFinalInstrumentation should only be called if a method body has been set for this function");
                 return E_FAIL;
@@ -512,6 +517,11 @@ namespace iast
         }
 
         FreeBuffer(); //To save memory
+
+        if (_applyMessage.size() > 0)
+        {
+            trace::Logger::Info(" --> ", _applyMessage, " <-- ");
+        }
         return hr;
     }
