+++ tracer/src/Datadog.Tracer.Native/iast/method_info.h
@@ -118,15 +118,15 @@ namespace iast
         bool _allowRestoreOnSecondJit = false;
         bool _disableInlining = false;
         bool _isWritten = false;
-        bool _isInstrumented = false;
 
         LPCBYTE _pOriginalMehodIL = nullptr;
-        ULONG _nOriginalMehodIL = 0;
+        DWORD _nOriginalMehodIL = 0;
         LPBYTE _pMethodIL = nullptr;
-        ULONG _nMethodIL = 0;
+        DWORD _nMethodIL = 0;
 
     protected:
         ILRewriter* _rewriter = nullptr;
+        std::string _applyMessage = "";
     private:
         void FreeBuffer();
     public:
@@ -139,16 +139,14 @@ namespace iast
         bool IsExcluded();
         bool IsProcessed();
         void SetProcessed();
-        bool IsInstrumented();
-        void SetInstrumented(bool instrumented);
-        bool HasChanges();
+        bool HasChanged();
         bool IsWritten();
         void DisableRestoreOnSecondJit();
         bool IsInlineEnabled();
         void DisableInlining();
 
-        HRESULT GetILRewriter(ILRewriter** rewriter, ICorProfilerInfo* pCorProfilerInfo = nullptr);
-        HRESULT CommitILRewriter(bool abort = false);
+        HRESULT GetILRewriter(ILRewriter** rewriter);
+        HRESULT CommitILRewriter(const std::string& applyMessage = "");
         HRESULT GetMethodIL(LPCBYTE* ppMehodIL, ULONG* pnSize, bool original = false);
         
         HRESULT SetMethodIL(ULONG nSize, LPCBYTE pMehodIL, ICorProfilerFunctionControl* pFunctionControl = nullptr);
@@ -156,6 +154,8 @@ namespace iast
 
         void ReJITCompilationStarted();
         void ReJITCompilationFinished();
+
+        void DumpIL(std::string message = "", ULONG pnMethodIL = 0, LPCBYTE pMethodIL = nullptr);
 
         bool IsPropertyAccessor();
         std::vector<WSTRING> GetCustomAttributes() override;
