+++ tracer/src/Datadog.Tracer.Native/iast/string_optimization_aspect_filter.cpp
@@ -1,5 +1,4 @@
 #include "string_optimization_aspect_filter.h"
-#include "dataflow_aspects.h"
 
 namespace iast
 {
@@ -23,11 +22,8 @@ namespace iast
         _targetMemberRefs.clear();
     }
 
-    bool StringOptimizationAspectFilter::AllowInstruction(DataflowContext& context)
+    bool StringOptimizationAspectFilter::AllowInstruction(ILInstr* instruction, ILRewriter* processor)
     {
-        ILInstr* instruction = context.instruction;
-        ILRewriter* processor = context.rewriter;
-
         if (_targetMemberRefs.empty()) { return true; }
         ILInstr* instr = instruction->m_pNext;
         // if the resulting string it is stored we cannot filter
