+++ tracer/src/Datadog.Tracer.Native/iast/string_optimization_aspect_filter.h
@@ -31,7 +31,7 @@ namespace iast
 		StringOptimizationAspectFilter(ModuleAspects* module);
 		~StringOptimizationAspectFilter() override;
 
-		bool AllowInstruction(DataflowContext& context) override;
+		bool AllowInstruction(ILInstr* instruction, ILRewriter* processor) override;
 	
 	};
 }
