+++ tracer/src/Datadog.Tracer.Native/method_rewriter.h
@@ -27,12 +27,35 @@ public:
     {        
     }
 
-    virtual HRESULT Rewrite(RejitHandlerModule* moduleHandler, RejitHandlerModuleMethod* methodHandler,
-                            ICorProfilerFunctionControl* pFunctionControl, ICorProfilerInfo* corProfilerInfo) = 0;
-    virtual InstrumentingProducts GetInstrumentingProduct(RejitHandlerModule* moduleHandler, RejitHandlerModuleMethod* methodHandler) = 0;
-    virtual WSTRING GetInstrumentationId(RejitHandlerModule* moduleHandler, RejitHandlerModuleMethod* methodHandler) = 0;
+    virtual HRESULT Rewrite(RejitHandlerModule* moduleHandler, RejitHandlerModuleMethod* methodHandler, ICorProfilerFunctionControl* pFunctionControl) = 0;
+    virtual InstrumentingProducts GetInstrumentingProduct(RejitHandlerModule* moduleHandler,
+                                              RejitHandlerModuleMethod* methodHandler) = 0;
+    virtual WSTRING GetInstrumentationId(RejitHandlerModule* moduleHandler,
+                                              RejitHandlerModuleMethod* methodHandler) = 0;
 
     virtual ~MethodRewriter() = default;
+};
+
+
+class TracerMethodRewriter : public MethodRewriter
+{
+private:
+    ILInstr* CreateFilterForException(ILRewriterWrapper* rewriter, mdTypeRef exception, mdTypeRef type_ref, mdMethodDef containsCallTargetBubbleUpException, ULONG exceptionValueIndex);
+
+public:
+
+    TracerMethodRewriter(CorProfiler* corProfiler) : MethodRewriter(corProfiler)
+    {
+    }
+
+    HRESULT Rewrite(RejitHandlerModule* moduleHandler, RejitHandlerModuleMethod* methodHandler, ICorProfilerFunctionControl* pFunctionControl) override;
+
+    InstrumentingProducts GetInstrumentingProduct(RejitHandlerModule* moduleHandler, RejitHandlerModuleMethod* methodHandler) override;
+    WSTRING GetInstrumentationId(RejitHandlerModule* moduleHandler, RejitHandlerModuleMethod* methodHandler) override;
+
+    std::tuple<WSTRING, WSTRING> GetResourceNameAndOperationName(const ComPtr<IMetaDataImport2>& metadataImport,
+                                                             const FunctionInfo* caller,
+                                                             TracerTokens* tracerTokens) const;
 };
 
 } // namespace trace
