+++ tracer/src/Datadog.Tracer.Native/rejit_handler.h
@@ -1,17 +1,19 @@
-#pragma once
+#ifndef DD_CLR_PROFILER_REJIT_HANDLER_H_
+#define DD_CLR_PROFILER_REJIT_HANDLER_H_
+
 #include <atomic>
-#include <future>
 #include <mutex>
 #include <shared_mutex>
 #include <string>
 #include <unordered_map>
 #include <vector>
+#include <future>
 
 #include "cor.h"
 #include "corprof.h"
-#include "method_rewriter.h"
 #include "module_metadata.h"
 #include "rejit_work_offloader.h"
+#include "method_rewriter.h"
 
 namespace trace
 {
@@ -40,10 +42,12 @@ protected:
     RejitHandlerModule* m_module;
 
 public:
-    RejitHandlerModuleMethod(mdMethodDef methodDef, RejitHandlerModule* module, const FunctionInfo& functionInfo,
-                             std::unique_ptr<MethodRewriter> methodRewriter);
+    RejitHandlerModuleMethod(mdMethodDef methodDef, RejitHandlerModule* module, const FunctionInfo& functionInfo, std::unique_ptr<MethodRewriter> methodRewriter); 
     mdMethodDef GetMethodDef();
     RejitHandlerModule* GetModule();
+
+    ICorProfilerFunctionControl* GetFunctionControl();
+    void SetFunctionControl(ICorProfilerFunctionControl* pFunctionControl);
 
     FunctionInfo* GetFunctionInfo();
     void SetFunctionInfo(const FunctionInfo& functionInfo);
@@ -54,8 +58,26 @@ public:
     virtual ~RejitHandlerModuleMethod() = default;
 };
 
-using RejitHandlerModuleMethodCreatorFunc =
-    std::function<std::unique_ptr<RejitHandlerModuleMethod>(const mdMethodDef, RejitHandlerModule*)>;
+/// <summary>
+/// Rejit handler representation of a method
+/// </summary>
+class TracerRejitHandlerModuleMethod : public RejitHandlerModuleMethod
+{
+private:
+    std::unique_ptr<IntegrationDefinition> m_integrationDefinition;
+
+public:
+    TracerRejitHandlerModuleMethod(
+                    mdMethodDef methodDef,
+                    RejitHandlerModule* module,
+                    const FunctionInfo& functionInfo,
+                    const IntegrationDefinition& integrationDefinition,
+                    std::unique_ptr<MethodRewriter> methodRewriter);
+    
+    IntegrationDefinition* GetIntegrationDefinition();
+};
+
+using RejitHandlerModuleMethodCreatorFunc = std::function<std::unique_ptr<RejitHandlerModuleMethod>(const mdMethodDef, RejitHandlerModule*)>;
 using RejitHandlerModuleMethodUpdaterFunc = std::function<void(RejitHandlerModuleMethod*)>;
 
 /// <summary>
@@ -68,7 +90,7 @@ private:
     std::unique_ptr<ModuleMetadata> m_metadata;
     std::mutex m_methods_lock;
     std::unordered_map<mdMethodDef, std::unique_ptr<RejitHandlerModuleMethod>> m_methods;
-
+    
     std::mutex m_ngenProcessedInlinerModulesLock;
     std::unordered_map<ModuleID, bool> m_ngenProcessedInlinerModules;
 
@@ -90,8 +112,6 @@ public:
     void RequestRejitForInlinersInModule(ModuleID moduleId);
 };
 
-class Rejitter;
-
 /// <summary>
 /// Class to control the ReJIT mechanism and to make sure all the required
 /// information is present before calling a method rewrite
@@ -102,32 +122,43 @@ private:
     std::atomic_bool m_shutdown = {false};
     Lock m_shutdown_lock;
 
+    std::mutex m_modules_lock;
+    std::unordered_map<ModuleID, std::unique_ptr<RejitHandlerModule>> m_modules;
     AssemblyProperty* m_pCorAssemblyProperty = nullptr;
 
     ICorProfilerInfo7* m_profilerInfo;
     ICorProfilerInfo10* m_profilerInfo10;
 
     std::shared_ptr<RejitWorkOffloader> m_work_offloader;
-
+        
     bool enable_by_ref_instrumentation = false;
     bool enable_calltarget_state_by_ref = false;
 
-    std::vector<Rejitter*> m_rejitters;
+    std::mutex m_ngenInlinersModules_lock;
+    std::vector<ModuleID> m_ngenInlinersModules;
 
 public:
     RejitHandler(ICorProfilerInfo7* pInfo, std::shared_ptr<RejitWorkOffloader> work_offloader);
     RejitHandler(ICorProfilerInfo10* pInfo, std::shared_ptr<RejitWorkOffloader> work_offloader);
 
-    void RegisterRejitter(Rejitter* rejitter);
-
+    RejitHandlerModule* GetOrAddModule(ModuleID moduleId);
     void SetEnableByRefInstrumentation(bool enableByRefInstrumentation);
     void SetEnableCallTargetStateByRef(bool enableCallTargetStateByRef);
     bool GetEnableCallTargetStateByRef();
     bool GetEnableByRefInstrumentation();
 
-    void RequestRejit(std::vector<ModuleID>& modulesVector, std::vector<mdMethodDef>& modulesMethodDef, bool callRevertExplicitly = false);
-    void EnqueueForRejit(std::vector<ModuleID>& modulesVector, std::vector<mdMethodDef>& modulesMethodDef, std::shared_ptr<std::promise<void>> promise = nullptr, bool callRevertExplicitly = false);
-    void EnqueueRequestRejit(std::vector<MethodIdentifier>& rejitRequests, std::shared_ptr<std::promise<void>> promise, bool callRevertExplicitly = false);
+    void RemoveModule(ModuleID moduleId);
+    bool HasModuleAndMethod(ModuleID moduleId, mdMethodDef methodDef);
+
+    void AddNGenInlinerModule(ModuleID moduleId);
+
+    void EnqueueForRejit(std::vector<ModuleID>& modulesVector, std::vector<mdMethodDef>& modulesMethodDef, std::shared_ptr<std::promise<void>> promise = nullptr);
+    void RequestRejit(std::vector<ModuleID>& modulesVector, std::vector<mdMethodDef>& modulesMethodDef);
+    void RequestRevert(std::vector<ModuleID>& modulesVector, std::vector<mdMethodDef>& modulesMethodDef);
+    void EnqueueForRevert(std::vector<ModuleID>& modulesVector, std::vector<mdMethodDef>& modulesMethodDef, std::shared_ptr<std::promise<void>> promise = nullptr);
+
+    void EnqueueRequestRejit(std::vector<MethodIdentifier>& rejitRequests, std::shared_ptr<std::promise<void>> promise);
+    void EnqueueRequestRevert(std::vector<MethodIdentifier>& revertRequests, std::shared_ptr<std::promise<void>> promise);
 
     void Shutdown();
     bool IsShutdownRequested();
@@ -140,10 +171,8 @@ public:
 
     void SetCorAssemblyProfiler(AssemblyProperty* pCorAssemblyProfiler);
     AssemblyProperty* GetCorAssemblyProperty();
-
-    bool HasModuleAndMethod(ModuleID moduleId, mdMethodDef methodDef);
-    void RemoveModule(ModuleID moduleId);
-    void AddNGenInlinerModule(ModuleID moduleId);
 };
 
 } // namespace trace
+
+#endif // DD_CLR_PROFILER_REJIT_HANDLER_H_
