+++ tracer/src/Datadog.Tracer.Native/rejit_preprocessor.cpp
@@ -1,274 +1,86 @@
 #include "rejit_preprocessor.h"
-#include "debugger_members.h"
-#include "fault_tolerant_tracker.h"
-#include "function_control_wrapper.h"
+#include "stats.h"
 #include "integration.h"
 #include "logger.h"
-#include "stats.h"
+#include "debugger_members.h"
+#include "fault_tolerant_tracker.h"
 
 namespace trace
 {
-// Rejitter
-
-Rejitter::Rejitter(std::shared_ptr<RejitHandler> handler, RejitterPriority priority) :
-    m_rejitHandler(handler), m_priority(priority)
-{
-    handler->RegisterRejitter(this);
-}
-
-Rejitter::~Rejitter()
-{
-}
 
 // RejitPreprocessor
 template <class RejitRequestDefinition>
 RejitPreprocessor<RejitRequestDefinition>::RejitPreprocessor(CorProfiler* corProfiler,
                                                              std::shared_ptr<RejitHandler> rejit_handler,
-                                                             std::shared_ptr<RejitWorkOffloader> work_offloader,
-                                                             RejitterPriority priority) :
-    Rejitter(rejit_handler, priority),
-    m_corProfiler(corProfiler),
-    m_rejit_handler(std::move(rejit_handler)),
-    m_work_offloader(std::move(work_offloader))
+                                                             std::shared_ptr<RejitWorkOffloader> work_offloader) :
+    m_corProfiler(corProfiler), m_rejit_handler(std::move(rejit_handler)), m_work_offloader(std::move(work_offloader))
 {
 }
 
 template <class RejitRequestDefinition>
-void RejitPreprocessor<RejitRequestDefinition>::Shutdown()
-{
-    Logger::Debug("RejitPreprocessor::Shutdown");
-
-    std::lock_guard<std::mutex> moduleGuard(m_modules_lock);
-    std::lock_guard<std::mutex> ngenModuleGuard(m_ngenInlinersModules_lock);
-
-    m_modules.clear();
-    m_ngenInlinersModules.clear();
-}
-
-template <class RejitRequestDefinition>
-RejitHandlerModule* RejitPreprocessor<RejitRequestDefinition>::GetOrAddModule(ModuleID moduleId)
-{
-    if (m_rejit_handler->IsShutdownRequested())
-    {
-        return nullptr;
-    }
-
-    std::lock_guard<std::mutex> guard(m_modules_lock);
-    auto find_res = m_modules.find(moduleId);
-    if (find_res != m_modules.end())
-    {
-        return find_res->second.get();
-    }
-
-    RejitHandlerModule* moduleHandler = new RejitHandlerModule(moduleId, m_rejit_handler.get());
-    m_modules[moduleId] = std::unique_ptr<RejitHandlerModule>(moduleHandler);
-    return moduleHandler;
-}
-
-template <class RejitRequestDefinition>
-bool RejitPreprocessor<RejitRequestDefinition>::HasModuleAndMethod(ModuleID moduleId, mdMethodDef methodDef)
-{
-    if (m_rejit_handler->IsShutdownRequested())
-    {
-        return false;
-    }
-
-    std::lock_guard<std::mutex> guard(m_modules_lock);
-    auto find_res = m_modules.find(moduleId);
-    if (find_res != m_modules.end())
-    {
-        auto moduleHandler = find_res->second.get();
-        return moduleHandler->ContainsMethod(methodDef);
-    }
-
-    return false;
-}
-
-template <class RejitRequestDefinition>
-void RejitPreprocessor<RejitRequestDefinition>::RemoveModule(ModuleID moduleId)
-{
-    if (m_rejit_handler->IsShutdownRequested())
-    {
-        return;
-    }
-
-    // Removes the RejitHandlerModule instance
-    std::lock_guard<std::mutex> modulesGuard(m_modules_lock);
-    m_modules.erase(moduleId);
-
-    // Removes the moduleID from the inliners vector
-    std::lock_guard<std::mutex> inlinersGuard(m_ngenInlinersModules_lock);
-    m_ngenInlinersModules.erase(std::remove(m_ngenInlinersModules.begin(), m_ngenInlinersModules.end(), moduleId),
-                                m_ngenInlinersModules.end());
-}
-
-template <class RejitRequestDefinition>
-void RejitPreprocessor<RejitRequestDefinition>::AddNGenInlinerModule(ModuleID moduleId)
-{
-    if (m_rejit_handler->IsShutdownRequested())
-    {
-        // If the shutdown was requested, we return.
-        return;
-    }
-
-    // Process the inliner module list ( to catch any incomplete data module )
-    // and also check if the module is already in the inliners list
-    std::lock_guard<std::mutex> modulesGuard(m_modules_lock);
-    std::lock_guard<std::mutex> inlinersGuard(m_ngenInlinersModules_lock);
-
-    bool alreadyAdded = false;
-    for (const auto& moduleInliner : m_ngenInlinersModules)
-    {
-        if (moduleInliner == moduleId)
-        {
-            alreadyAdded = true;
-        }
-
-        for (const auto& mod : m_modules)
-        {
-            mod.second->RequestRejitForInlinersInModule(moduleInliner);
-        }
-    }
-
-    // If the module is not in the inliners list we added and request rejit for it.
-    if (!alreadyAdded)
-    {
-        // Add the new module inliner
-        m_ngenInlinersModules.push_back(moduleId);
-
-        for (const auto& mod : m_modules)
-        {
-            mod.second->RequestRejitForInlinersInModule(moduleId);
-        }
-    }
-}
-
-template <class RejitRequestDefinition>
-HRESULT RejitPreprocessor<RejitRequestDefinition>::RejitMethod(FunctionControlWrapper& functionControl)
-{
-    auto moduleId = functionControl.GetModuleId();
-    auto methodId = functionControl.GetMethodId();
-
-    auto moduleHandler = GetOrAddModule(moduleId);
-    if (moduleHandler == nullptr)
-    {
-        return S_FALSE;
-    }
-
-    RejitHandlerModuleMethod* methodHandler = nullptr;
-    if (!moduleHandler->TryGetMethod(methodId, &methodHandler))
-    {
-        return S_FALSE;
-    }
-
-    if (methodHandler->GetMethodDef() == mdMethodDefNil)
-    {
-        Logger::Warn("RejitMethod: mdMethodDef is missing for "
-                     "MethodDef: ",
-                     methodId);
-        return S_FALSE;
-    }
-
-    if (methodHandler->GetFunctionInfo() == nullptr)
-    {
-        Logger::Warn("RejitMethod: FunctionInfo is missing for "
-                     "MethodDef: ",
-                     methodId);
-        return S_FALSE;
-    }
-
-    if (moduleHandler->GetModuleId() == 0)
-    {
-        Logger::Warn("RejitMethod: ModuleID is missing for "
-                     "MethodDef: ",
-                     methodId);
-        return S_FALSE;
-    }
-
-    if (moduleHandler->GetModuleMetadata() == nullptr)
-    {
-        Logger::Warn("RejitMethod: ModuleMetadata is missing for "
-                     "MethodDef: ",
-                     methodId);
-        return S_FALSE;
-    }
-
-    auto rewriter = methodHandler->GetMethodRewriter();
-
-    if (rewriter == nullptr)
-    {
-        Logger::Error("RejitMethod: The rewriter is missing for "
-                      "MethodDef: ",
-                      methodId, ", methodHandler type name = ", typeid(methodHandler).name());
-        return S_FALSE;
-    }
-
-    return rewriter->Rewrite(moduleHandler, methodHandler, (ICorProfilerFunctionControl*) &functionControl,
-                             (ICorProfilerInfo*) &functionControl);
-}
-
-template <class RejitRequestDefinition>
-void RejitPreprocessor<RejitRequestDefinition>::EnqueueFaultTolerantMethods(
-    const RejitRequestDefinition& definition, ComPtr<IMetaDataImport2>& metadataImport,
-    ComPtr<IMetaDataEmit2>& metadataEmit, const ModuleInfo& moduleInfo, const mdTypeDef typeDef,
-    std::vector<MethodIdentifier>& rejitRequests, unsigned methodDef, const FunctionInfo& functionInfo,
-    RejitHandlerModule* moduleHandler)
+void RejitPreprocessor<RejitRequestDefinition>::EnqueueFaultTolerantMethods(const RejitRequestDefinition& definition, ComPtr<IMetaDataImport2>& metadataImport, ComPtr<IMetaDataEmit2>& metadataEmit, const ModuleInfo& moduleInfo, const mdTypeDef typeDef, std::vector<MethodIdentifier>& rejitRequests, unsigned methodDef, const FunctionInfo& functionInfo, RejitHandlerModule* moduleHandler)
 {
     if (fault_tolerant::FaultTolerantTracker::Instance()->IsKickoffMethod(moduleInfo.id, methodDef))
     {
         const auto originalMethod =
             fault_tolerant::FaultTolerantTracker::Instance()->GetOriginalMethod(moduleInfo.id, methodDef);
-        RejitPreprocessor::EnqueueNewMethod(definition, metadataImport, metadataEmit, moduleInfo, typeDef,
-                                            rejitRequests, originalMethod, functionInfo, moduleHandler);
+        RejitPreprocessor::EnqueueNewMethod(definition, metadataImport, metadataEmit, moduleInfo, typeDef, rejitRequests,
+                                            originalMethod,
+                                            functionInfo, moduleHandler);
 
         const auto instrumentedMethod =
             fault_tolerant::FaultTolerantTracker::Instance()->GetInstrumentedMethod(moduleInfo.id, methodDef);
         RejitPreprocessor::EnqueueNewMethod(definition, metadataImport, metadataEmit, moduleInfo, typeDef,
-                                            rejitRequests, instrumentedMethod, functionInfo, moduleHandler);
+                                            rejitRequests,
+                                            instrumentedMethod,
+                                            functionInfo, moduleHandler);
     }
 }
 
 template <class RejitRequestDefinition>
-void RejitPreprocessor<RejitRequestDefinition>::EnqueueNewMethod(const RejitRequestDefinition& definition,
-                                                                 ComPtr<IMetaDataImport2>& metadataImport,
-                                                                 ComPtr<IMetaDataEmit2>& metadataEmit,
-                                                                 const ModuleInfo& moduleInfo, const mdTypeDef typeDef,
-                                                                 std::vector<MethodIdentifier>& rejitRequests,
-                                                                 unsigned methodDef, const FunctionInfo& functionInfo,
-                                                                 RejitHandlerModule* moduleHandler)
+void RejitPreprocessor<RejitRequestDefinition>::EnqueueNewMethod(const RejitRequestDefinition& definition, 
+    ComPtr<IMetaDataImport2>& metadataImport, 
+    ComPtr<IMetaDataEmit2>& metadataEmit, 
+    const ModuleInfo& moduleInfo, 
+    const mdTypeDef typeDef, 
+    std::vector<MethodIdentifier>& rejitRequests, 
+    unsigned methodDef,
+    const FunctionInfo& functionInfo, 
+    RejitHandlerModule* moduleHandler)
 {
     EnqueueFaultTolerantMethods(definition, metadataImport, metadataEmit, moduleInfo, typeDef, rejitRequests, methodDef,
-                                functionInfo, moduleHandler);
+                               functionInfo, moduleHandler);
 
-    RejitHandlerModuleMethodCreatorFunc creator =
-        [=, request = definition, fInfo = functionInfo](const mdMethodDef method, RejitHandlerModule* module) {
-            return CreateMethod(method, module, fInfo, request);
-        };
+    RejitHandlerModuleMethodCreatorFunc creator = [=, request = definition, fInfo = functionInfo](
+        const mdMethodDef method, RejitHandlerModule* module) {
+        return CreateMethod(method, module, fInfo, request);
+    };
 
     RejitHandlerModuleMethodUpdaterFunc updater = [=, request = definition](RejitHandlerModuleMethod* method) {
         return UpdateMethod(method, request);
     };
 
     moduleHandler->CreateMethodIfNotExists(methodDef, creator, updater);
-
+    
     // Store module_id and methodDef to request the ReJIT after analyzing all integrations.
     rejitRequests.emplace_back(MethodIdentifier(moduleInfo.id, methodDef));
 }
 
 template <class RejitRequestDefinition>
-void RejitPreprocessor<RejitRequestDefinition>::ProcessTypeDefForRejit(
-    const RejitRequestDefinition& definition, ComPtr<IMetaDataImport2>& metadataImport,
-    ComPtr<IMetaDataEmit2>& metadataEmit, ComPtr<IMetaDataAssemblyImport>& assemblyImport,
-    ComPtr<IMetaDataAssemblyEmit>& assemblyEmit, const ModuleInfo& moduleInfo, const mdTypeDef typeDef,
-    std::vector<MethodIdentifier>& rejitRequests)
+void RejitPreprocessor<RejitRequestDefinition>::ProcessTypeDefForRejit(const RejitRequestDefinition& definition,
+                                          ComPtr<IMetaDataImport2>& metadataImport,
+                                          ComPtr<IMetaDataEmit2>& metadataEmit,
+                                          ComPtr<IMetaDataAssemblyImport>& assemblyImport,
+                                          ComPtr<IMetaDataAssemblyEmit>& assemblyEmit, const ModuleInfo& moduleInfo,
+                                          const mdTypeDef typeDef, std::vector<MethodIdentifier>& rejitRequests)
 {
     auto target_method = GetTargetMethod(definition);
     auto is_interface = GetIsInterface(definition);
     const bool wildcard_enabled = target_method.method_name == tracemethodintegration_wildcardmethodname;
     const bool iterate_explicit_interface_methods = is_interface && !wildcard_enabled;
 
-    Logger::Debug("  Looking for '", target_method.type.name, ".", target_method.method_name, "(",
-                  (target_method.signature_types.size() - 1), " params)' method implementation.");
+    Logger::Debug("  Looking for '", target_method.type.name, ".", target_method.method_name,
+                  "(", (target_method.signature_types.size() - 1), " params)' method implementation.");
 
     // Now we enumerate all methods with the same target method name. (All overloads of the method)
     auto enumMethods = Enumerator<mdMethodDef>(
@@ -279,15 +91,16 @@ void RejitPreprocessor<RejitRequestDefinition>::ProcessTypeDefForRejit(
             }
             else
             {
-                return metadataImport->EnumMethodsWithName(ptr, typeDef, target_method.method_name.c_str(), arr, max,
-                                                           cnt);
+                return metadataImport->EnumMethodsWithName(ptr, typeDef, target_method.method_name.c_str(), arr,
+                                                           max, cnt);
             }
         },
         [&metadataImport](HCORENUM ptr) -> void { metadataImport->CloseEnum(ptr); });
     auto enumExplicitInterfaceMethods = Enumerator<mdMethodDef>(
         [&metadataImport, target_method, typeDef](HCORENUM* ptr, mdMethodDef arr[], ULONG max, ULONG* cnt) -> HRESULT {
             auto method_name = target_method.type.name + WStr(".") + target_method.method_name;
-            return metadataImport->EnumMethodsWithName(ptr, typeDef, method_name.c_str(), arr, max, cnt);
+            return metadataImport->EnumMethodsWithName(ptr, typeDef, method_name.c_str(), arr,
+                                                        max, cnt);
         },
         [&metadataImport](HCORENUM ptr) -> void { metadataImport->CloseEnum(ptr); });
 
@@ -296,7 +109,9 @@ void RejitPreprocessor<RejitRequestDefinition>::ProcessTypeDefForRejit(
     auto enable_by_ref_instrumentation = m_rejit_handler->GetEnableByRefInstrumentation();
     auto enable_calltarget_state_by_ref = m_rejit_handler->GetEnableCallTargetStateByRef();
 
-    auto enumIterator = iterate_explicit_interface_methods ? enumExplicitInterfaceMethods.begin() : enumMethods.begin();
+    auto enumIterator =
+        iterate_explicit_interface_methods ?
+            enumExplicitInterfaceMethods.begin() : enumMethods.begin();
     auto iteratorEnd = enumMethods.end();
     auto explicitMode = iterate_explicit_interface_methods;
 
@@ -333,22 +148,20 @@ void RejitPreprocessor<RejitRequestDefinition>::ProcessTypeDefForRejit(
         if (FAILED(hr))
         {
             Logger::Warn("    * Skipping ", functionInfo.method_signature.str(),
-                         ": the method signature cannot be parsed.");
+                ": the method signature cannot be parsed.");
             continue;
         }
 
         const auto numOfArgs = functionInfo.method_signature.NumberOfArguments();
         if (wildcard_enabled)
         {
-            if (tracemethodintegration_wildcard_ignored_methods.find(caller.name) !=
-                    tracemethodintegration_wildcard_ignored_methods.end() ||
+            if (tracemethodintegration_wildcard_ignored_methods.find(caller.name) != tracemethodintegration_wildcard_ignored_methods.end() ||
                 caller.name.find(tracemethodintegration_setterprefix) == 0 ||
                 caller.name.find(tracemethodintegration_getterprefix) == 0)
             {
                 Logger::Warn(
-                    "    * Skipping enqueue for ReJIT, special method detected during '*' wildcard search [ModuleId=",
-                    moduleInfo.id, ", MethodDef=", shared::TokenStr(&methodDef), ", Type=", caller.type.name,
-                    ", Method=", caller.name, "(", numOfArgs, " params), Signature=", caller.signature.str(), "]");
+                    "    * Skipping enqueue for ReJIT, special method detected during '*' wildcard search [ModuleId=", moduleInfo.id, ", MethodDef=", shared::TokenStr(&methodDef),
+                    ", Type=", caller.type.name, ", Method=", caller.name, "(", numOfArgs, " params), Signature=", caller.signature.str(), "]");
                 continue;
             }
         }
@@ -361,7 +174,7 @@ void RejitPreprocessor<RejitRequestDefinition>::ProcessTypeDefForRejit(
             if (numOfArgs != target_method.signature_types.size() - 1)
             {
                 Logger::Info("    * Skipping ", caller.type.name, ".", caller.name,
-                             ": the methoddef doesn't have the right number of arguments (", numOfArgs, " arguments).");
+                    ": the methoddef doesn't have the right number of arguments (", numOfArgs, " arguments).");
                 continue;
             }
 
@@ -384,14 +197,14 @@ void RejitPreprocessor<RejitRequestDefinition>::ProcessTypeDefForRejit(
             if (argumentsMismatch)
             {
                 Logger::Info("    * Skipping ", target_method.method_name,
-                             ": the methoddef doesn't have the right type of arguments.");
+                    ": the methoddef doesn't have the right type of arguments.");
                 continue;
             }
         }
 
         // As we are in the right method, we gather all information we need and stored it in to the
         // ReJIT handler.
-        auto moduleHandler = GetOrAddModule(moduleInfo.id);
+        auto moduleHandler = m_rejit_handler->GetOrAddModule(moduleInfo.id);
         if (moduleHandler == nullptr)
         {
             Logger::Warn("Module handler is null, this only happens if the RejitHandler has been shutdown.");
@@ -407,22 +220,20 @@ void RejitPreprocessor<RejitRequestDefinition>::ProcessTypeDefForRejit(
                                    enable_by_ref_instrumentation, enable_calltarget_state_by_ref);
 
             Logger::Debug("ReJIT handler stored metadata for ", moduleInfo.id, " ", moduleInfo.assembly.name,
-                          " AppDomain ", moduleInfo.assembly.app_domain_id, " ", moduleInfo.assembly.app_domain_name);
+                         " AppDomain ", moduleInfo.assembly.app_domain_id, " ", moduleInfo.assembly.app_domain_name);
 
             moduleHandler->SetModuleMetadata(moduleMetadata);
         }
 
-        Logger::Info("Method enqueued for ReJIT for ", target_method.type.name, ".", target_method.method_name, "(",
-                     (target_method.signature_types.size() - 1), " params).");
+        Logger::Info("Method enqueued for ReJIT for ", target_method.type.name, ".", target_method.method_name,
+                  "(", (target_method.signature_types.size() - 1), " params).");
         EnqueueNewMethod(definition, metadataImport, metadataEmit, moduleInfo, typeDef, rejitRequests, methodDef,
                          functionInfo, moduleHandler);
 
-        // If we are in the explicit enumerator and we found the method, we don't look into the normal methods
-        // enumerators.
+        // If we are in the explicit enumerator and we found the method, we don't look into the normal methods enumerators.
         if (explicitMode)
         {
-            Logger::Debug("      Explicit interface implementation found, skipping normal methods search. [",
-                          caller.type.name, ".", caller.name, "]");
+            Logger::Debug("      Explicit interface implementation found, skipping normal methods search. [", caller.type.name, ".", caller.name, "]");
             break;
         }
     }
@@ -452,18 +263,29 @@ void RejitPreprocessor<RejitRequestDefinition>::ProcessTypesForRejit(
 
 template <class RejitRequestDefinition>
 ULONG RejitPreprocessor<RejitRequestDefinition>::RequestRejitForLoadedModules(
-    const std::vector<ModuleID>& modules, const std::vector<RejitRequestDefinition>& definitions,
-    bool enqueueInSameThread)
+                                                        const std::vector<ModuleID>& modules,
+                                                        const std::vector<RejitRequestDefinition>& definitions,
+                                                        bool enqueueInSameThread)
 {
-    std::vector<MethodIdentifier> rejitRequests{};
-    const auto rejitCount = PreprocessRejitRequests(modules, definitions, rejitRequests);
+    std::vector<MethodIdentifier> rejitRequests {};
+    const auto rejitCount = PreprocessRejitRequests(modules, definitions, rejitRequests, false);
     RequestRejit(rejitRequests, enqueueInSameThread);
     return rejitCount;
 }
 
 template <class RejitRequestDefinition>
-void RejitPreprocessor<RejitRequestDefinition>::RequestRejit(std::vector<MethodIdentifier>& rejitRequests,
-                                                             bool enqueueInSameThread, bool callRevertExplicitly)
+ULONG RejitPreprocessor<RejitRequestDefinition>::RequestRevertForLoadedModules(
+    const std::vector<ModuleID>& modules, const std::vector<RejitRequestDefinition>& definitions,
+    bool enqueueInSameThread)
+{
+    std::vector<MethodIdentifier> rejitRequests{};
+    const auto rejitCount = PreprocessRejitRequests(modules, definitions, rejitRequests, true);
+    RequestRevert(rejitRequests, enqueueInSameThread);
+    return rejitCount;
+}
+
+template <class RejitRequestDefinition>
+void RejitPreprocessor<RejitRequestDefinition>::RequestRejit(std::vector<MethodIdentifier>& rejitRequests, bool enqueueInSameThread)
 {
     if (!rejitRequests.empty())
     {
@@ -482,19 +304,47 @@ void RejitPreprocessor<RejitRequestDefinition>::RequestRejit(std::vector<MethodI
 
         if (enqueueInSameThread)
         {
-            m_rejit_handler->RequestRejit(vtModules, vtMethodDefs, callRevertExplicitly);
+            m_rejit_handler->RequestRejit(vtModules, vtMethodDefs);
         }
         else
         {
-            m_rejit_handler->EnqueueForRejit(vtModules, vtMethodDefs, nullptr, callRevertExplicitly);
+            m_rejit_handler->EnqueueForRejit(vtModules, vtMethodDefs);
+        }
+    }
+}
+
+template <class RejitRequestDefinition>
+void RejitPreprocessor<RejitRequestDefinition>::RequestRevert(std::vector<MethodIdentifier>& revertRequests, bool enqueueInSameThread)
+{
+    if (!revertRequests.empty())
+    {
+        std::vector<ModuleID> vtModules;
+        std::vector<mdMethodDef> vtMethodDefs;
+
+        const auto rejitCount = revertRequests.size();
+        vtModules.reserve(rejitCount);
+        vtMethodDefs.reserve(rejitCount);
+
+        for (const auto& rejitRequest : revertRequests)
+        {
+            vtModules.push_back(rejitRequest.moduleId);
+            vtMethodDefs.push_back(rejitRequest.methodToken);
+        }
+
+        if (enqueueInSameThread)
+        {
+            m_rejit_handler->RequestRevert(vtModules, vtMethodDefs);
+        }
+        else
+        {
+            m_rejit_handler->EnqueueForRevert(vtModules, vtMethodDefs);
         }
     }
 }
 
 template <class RejitRequestDefinition>
 void RejitPreprocessor<RejitRequestDefinition>::EnqueueRequestRejit(std::vector<MethodIdentifier>& rejitRequests,
-                                                                    std::shared_ptr<std::promise<void>> promise,
-                                                                    bool callRevertExplicitly)
+    std::shared_ptr<std::promise<void>> promise)
 {
     if (m_rejit_handler->IsShutdownRequested())
     {
@@ -513,9 +363,45 @@ void RejitPreprocessor<RejitRequestDefinition>::EnqueueRequestRejit(std::vector<
 
     Logger::Debug("RejitHandler::EnqueueRequestRejit");
 
-    std::function<void()> action = [=, requests = std::move(rejitRequests), localPromise = promise, callRevertExplicitly = callRevertExplicitly]() mutable {
+    std::function<void()> action = [=, requests = std::move(rejitRequests), localPromise = promise]() mutable {
         // Process modules for rejit
-        RequestRejit(requests, true, callRevertExplicitly);
+        RequestRejit(requests, true);
+
+        // Resolve promise
+        if (localPromise != nullptr)
+        {
+            localPromise->set_value();
+        }
+    };
+
+    // Enqueue
+    m_work_offloader->Enqueue(std::make_unique<RejitWorkItem>(std::move(action)));
+}
+
+template <class RejitRequestDefinition>
+void RejitPreprocessor<RejitRequestDefinition>::EnqueueRequestRevert(std::vector<MethodIdentifier>& revertRequests,
+                                                                    std::shared_ptr<std::promise<void>> promise)
+{
+    if (m_rejit_handler->IsShutdownRequested())
+    {
+        if (promise != nullptr)
+        {
+            promise->set_value();
+        }
+
+        return;
+    }
+
+    if (revertRequests.size() == 0)
+    {
+        return;
+    }
+
+    Logger::Debug("RejitHandler::EnqueueRequestRevert");
+
+    std::function<void()> action = [=, requests = std::move(revertRequests), localPromise = promise]() mutable {
+        // Process modules for rejit
+        RequestRevert(requests, true);
 
         // Resolve promise
         if (localPromise != nullptr)
@@ -567,9 +453,47 @@ void RejitPreprocessor<RejitRequestDefinition>::EnqueueRequestRejitForLoadedModu
 }
 
 template <class RejitRequestDefinition>
+void RejitPreprocessor<RejitRequestDefinition>::EnqueueRequestRevertForLoadedModules(
+    const std::vector<ModuleID>& modulesVector, const std::vector<RejitRequestDefinition>& definitions,
+    std::shared_ptr<std::promise<ULONG>> promise)
+{
+    if (m_rejit_handler->IsShutdownRequested())
+    {
+        if (promise != nullptr)
+        {
+            promise->set_value(0);
+        }
+
+        return;
+    }
+
+    if (modulesVector.size() == 0 || definitions.size() == 0)
+    {
+        return;
+    }
+
+    Logger::Debug("RejitHandler::EnqueueRequestRevertForLoadedModules");
+
+    std::function<void()> action = [=, modules = std::move(modulesVector), definitions = std::move(definitions),
+                                    localPromise = promise]() mutable {
+        // Process modules for rejit
+        const auto rejitCount = RequestRevertForLoadedModules(modules, definitions, true);
+
+        // Resolve promise
+        if (localPromise != nullptr)
+        {
+            localPromise->set_value(rejitCount);
+        }
+    };
+
+    // Enqueue
+    m_work_offloader->Enqueue(std::make_unique<RejitWorkItem>(std::move(action)));
+}
+
+template <class RejitRequestDefinition>
 ULONG RejitPreprocessor<RejitRequestDefinition>::PreprocessRejitRequests(
     const std::vector<ModuleID>& modules, const std::vector<RejitRequestDefinition>& definitions,
-    std::vector<MethodIdentifier>& rejitRequests)
+    std::vector<MethodIdentifier>& rejitRequests, bool isRevert)
 {
     if (m_rejit_handler->IsShutdownRequested())
     {
@@ -597,6 +521,18 @@ ULONG RejitPreprocessor<RejitRequestDefinition>::PreprocessRejitRequests(
             const auto is_derived = GetIsDerived(definition);
             const auto is_interface = GetIsInterface(definition);
             const auto is_enabled = GetIsEnabled(definition);
+
+            if (SupportsSelectiveEnablement())
+            {
+                if (!isRevert && !is_enabled)
+                {
+                    continue; // Disabled calltarget
+                }
+                else if (isRevert && is_enabled)
+                {
+                    continue; // Enabled calltarget
+                }
+            }
 
             if (is_derived || is_interface)
             {
@@ -675,13 +611,12 @@ ULONG RejitPreprocessor<RejitRequestDefinition>::PreprocessRejitRequests(
                         auto interfaceImplEnum = EnumInterfaceImpls(metadataImport, typeDef);
                         auto interfaceImplIterator = interfaceImplEnum.begin();
                         for (; interfaceImplIterator != interfaceImplEnum.end();
-                             interfaceImplIterator = ++interfaceImplIterator)
+                            interfaceImplIterator = ++interfaceImplIterator)
                         {
                             // Get the interface token
                             auto interfaceImpl = *interfaceImplIterator;
                             mdToken classToken, interfaceToken;
-                            if (metadataImport->GetInterfaceImplProps(interfaceImpl, &classToken, &interfaceToken) ==
-                                S_OK)
+                            if (metadataImport->GetInterfaceImplProps(interfaceImpl, &classToken, &interfaceToken) == S_OK)
                             {
                                 if (classToken == typeDef)
                                 {
@@ -693,9 +628,8 @@ ULONG RejitPreprocessor<RejitRequestDefinition>::PreprocessRejitRequests(
                                     if (interfaceTokenType == mdtTypeRef)
                                     {
                                         mdAssembly assemblyToken;
-                                        if (metadataImport->GetTypeRefProps(interfaceToken, &assemblyToken, type_name,
-                                                                            kNameMaxSize, &type_name_len) == S_OK &&
-                                            type_name == target_method.type.name)
+                                        if (metadataImport->GetTypeRefProps(interfaceToken, &assemblyToken, type_name, kNameMaxSize, &type_name_len) == S_OK
+                                            && type_name == target_method.type.name)
                                         {
                                             // Now we have to validate the assembly version
                                             const auto tokenType = TypeFromToken(assemblyToken);
@@ -724,10 +658,8 @@ ULONG RejitPreprocessor<RejitRequestDefinition>::PreprocessRejitRequests(
                                     {
                                         DWORD type_flags;
                                         mdToken type_extends = mdTokenNil;
-                                        if (metadataImport->GetTypeDefProps(interfaceToken, type_name, kNameMaxSize,
-                                                                            &type_name_len, &type_flags,
-                                                                            &type_extends) == S_OK &&
-                                            type_name == target_method.type.name)
+                                        if (metadataImport->GetTypeDefProps(interfaceToken, type_name, kNameMaxSize, &type_name_len, &type_flags, &type_extends) == S_OK
+                                            && type_name == target_method.type.name)
                                         {
                                             // We check the assembly name and version
                                             if (assemblyMetadata->name == target_method.type.assembly.name &&
@@ -857,8 +789,7 @@ ULONG RejitPreprocessor<RejitRequestDefinition>::PreprocessRejitRequests(
                     continue;
                 }
 
-                ProcessTypesForRejit(rejitRequests, moduleInfo, metadataImport, metadataEmit, assemblyImport,
-                                     assemblyEmit, definition, target_method);
+                ProcessTypesForRejit(rejitRequests, moduleInfo, metadataImport, metadataEmit, assemblyImport, assemblyEmit, definition, target_method);
             }
         }
     }
@@ -869,8 +800,7 @@ ULONG RejitPreprocessor<RejitRequestDefinition>::PreprocessRejitRequests(
 }
 
 template <class RejitRequestDefinition>
-void RejitPreprocessor<RejitRequestDefinition>::EnqueuePreprocessRejitRequests(
-    const std::vector<ModuleID>& modulesVector, const std::vector<RejitRequestDefinition>& definitions,
+void RejitPreprocessor<RejitRequestDefinition>::EnqueuePreprocessRejitRequests(const std::vector<ModuleID>& modulesVector, const std::vector<RejitRequestDefinition>& definitions,
     std::shared_ptr<std::promise<std::vector<MethodIdentifier>>> promise)
 {
     std::vector<MethodIdentifier> rejitRequests;
@@ -893,9 +823,11 @@ void RejitPreprocessor<RejitRequestDefinition>::EnqueuePreprocessRejitRequests(
     Logger::Debug("RejitHandler::EnqueuePreprocessRejitRequests");
 
     std::function<void()> action = [=, modules = std::move(modulesVector), definitions = std::move(definitions),
-                                    localRejitRequests = rejitRequests, localPromise = promise]() mutable {
+                                    localRejitRequests = rejitRequests,
+                                    localPromise = promise]() mutable {
+
         // Process modules for rejit
-        const auto rejitCount = PreprocessRejitRequests(modules, definitions, localRejitRequests);
+        const auto rejitCount = PreprocessRejitRequests(modules, definitions, localRejitRequests, true);
 
         // Resolve promise
         if (localPromise != nullptr)
@@ -912,6 +844,60 @@ template <class RejitRequestDefinition>
 void RejitPreprocessor<RejitRequestDefinition>::UpdateMethod(RejitHandlerModuleMethod* method,
                                                              const RejitRequestDefinition& definition)
 {
+}
+
+// TraceIntegrationRejitPreprocessor
+
+const MethodReference& TracerRejitPreprocessor::GetTargetMethod(const IntegrationDefinition& integrationDefinition)
+{
+    return integrationDefinition.target_method;
+}
+
+const bool TracerRejitPreprocessor::GetIsDerived(const IntegrationDefinition& integrationDefinition)
+{
+    return integrationDefinition.is_derived;
+}
+
+const bool TracerRejitPreprocessor::GetIsInterface(const IntegrationDefinition& integrationDefinition)
+{
+    return integrationDefinition.is_interface;
+}
+
+const bool TracerRejitPreprocessor::GetIsExactSignatureMatch(const IntegrationDefinition& integrationDefinition)
+{
+    return integrationDefinition.is_exact_signature_match;
+}
+
+const bool TracerRejitPreprocessor::GetIsEnabled(const IntegrationDefinition& integrationDefinition)
+{
+    return integrationDefinition.GetEnabled();
+}
+
+const bool TracerRejitPreprocessor::SupportsSelectiveEnablement()
+{
+    return true;
+}
+
+
+const std::unique_ptr<RejitHandlerModuleMethod> TracerRejitPreprocessor::CreateMethod(const mdMethodDef methodDef, RejitHandlerModule* module,
+                                                const FunctionInfo& functionInfo,
+                                                const IntegrationDefinition& integrationDefinition)
+{
+    return std::make_unique<TracerRejitHandlerModuleMethod>(methodDef,
+                                                            module,
+                                                            functionInfo,
+                                                            integrationDefinition,
+                                                            std::make_unique<TracerMethodRewriter>(m_corProfiler));
+}
+
+bool TracerRejitPreprocessor::ShouldSkipModule(const ModuleInfo& moduleInfo, const IntegrationDefinition& integrationDefinition)
+{
+    // If the integration is not for the current assembly we skip.
+
+    const auto target_method = GetTargetMethod(integrationDefinition);
+
+    return target_method.type.assembly.name != tracemethodintegration_assemblyname &&
+           target_method.type.assembly.name != moduleInfo.assembly.name;
 }
 
 template class RejitPreprocessor<std::shared_ptr<debugger::MethodProbeDefinition>>;
