+++ tracer/test/Datadog.Tracer.Native.Tests/integration_test.cpp
@@ -1,7 +1,6 @@
 #include "pch.h"
 
 #include "../../src/Datadog.Tracer.Native/integration.h"
-#include "../../src/Datadog.Tracer.Native/tracer_integration_definition.h"
 
 using namespace trace;
