+++ tracer/test/snapshots/Iast.CustomAttribute.AspNetCore5.IastEnabled.verified.txt
@@ -1,92 +1,1 @@
-﻿[
-  {
-    TraceId: Id_1,
-    SpanId: Id_2,
-    Name: aspnet_core.request,
-    Resource: GET /iast/customattribute,
-    Service: Samples.Security.AspNetCore5,
-    Type: web,
-    Tags: {
-      aspnet_core.endpoint: Samples.Security.AspNetCore5.Controllers.IastController.CustomAttribute (Samples.Security.AspNetCore5),
-      aspnet_core.route: iast/customattribute,
-      component: aspnet_core,
-      env: integration_tests,
-      http.method: GET,
-      http.request.headers.host: localhost:00000,
-      http.route: iast/customattribute,
-      http.status_code: 200,
-      http.url: http://localhost:00000/Iast/CustomAttribute?userName=Vicent,
-      http.useragent: Mistake Not...,
-      language: dotnet,
-      runtime-id: Guid_1,
-      span.kind: server,
-      _dd.iast.enabled: 1,
-      _dd.iast.json:
-{
-  "vulnerabilities": [
-    {
-      "type": "PATH_TRAVERSAL",
-      "hash": -2017497883,
-      "location": {
-        "spanId": XXX,
-        "path": "Samples.Security.AspNetCore5.Controllers.IastController",
-        "method": "GetCustomString"
-      },
-      "evidence": {
-        "valueParts": [
-          {
-            "value": "Vicent",
-            "source": 0
-          }
-        ]
-      }
-    }
-  ],
-  "sources": [
-    {
-      "origin": "http.request.parameter",
-      "name": "userName",
-      "value": "Vicent"
-    }
-  ]
-}
-    },
-    Metrics: {
-      process_id: 0,
-      _dd.top_level: 1.0,
-      _dd.tracer_kr: 1.0,
-      _sampling_priority_v1: 1.0
-    }
-  },
-  {
-    TraceId: Id_1,
-    SpanId: Id_3,
-    Name: aspnet_core_mvc.request,
-    Resource: GET /iast/customattribute,
-    Service: Samples.Security.AspNetCore5,
-    Type: web,
-    ParentId: Id_2,
-    Tags: {
-      aspnet_core.action: customattribute,
-      aspnet_core.controller: iast,
-      aspnet_core.route: iast/customattribute,
-      component: aspnet_core,
-      env: integration_tests,
-      language: dotnet,
-      span.kind: server
-    }
-  },
-  {
-    TraceId: Id_1,
-    SpanId: Id_4,
-    Name: span.custom.attribute,
-    Resource: IastController.GetCustomString,
-    Service: Samples.Security.AspNetCore5,
-    ParentId: Id_3,
-    Tags: {
-      component: trace,
-      env: integration_tests,
-      language: dotnet
-    }
-  }
-]
+﻿
